feat: add hideDivider prop to NavList
